requirements.txt
Django==1.10.2
Pillow
MySQL-python
django-crispy-forms==1.6.0
python-dateutil
django-hvad
django-localeurl
pytz
